fix: log-saturating marathon correction, move module to js/estimator
acceptance fixes (user review):
- order inversion fix: linear duration penalty flipped REFORM 2nd 8th
  (9.28) over 9th (9.25) because penalty difference (47s x 0.20/60)
  exceeded the 0.12 base gap. Switch to log-saturating correction
  corr = min(0.50, 0.40*ln(1+excessMin)) * taper - sublinear growth keeps
  adjacent dan-tier course charts ordered (8th 9.28 < 9th 9.32);
  order-constrained course grid: zero new inversions at every scale/cap
- calibration (user-authorized, course subset, order check added):
  scale=0.40 cap=0.50 -> merged MAE 0.3167 (vs linear 0.3225),
  Roxy MAE 0.4036->0.2250 (Exact 50%), Azusa MAE 0.4782->0.3544
  (Bias -0.424->-0.085)
- module moved js/rework/ -> js/estimator/marathonCorrection.js
  (acceptance: not a rework module); imports and docs updated
- docs: features/marathon-correction.md (log formula + dual-round
  calibration record), azusa_algorithm.md/roxy_algorithm.md sections,
  analysis-pipeline.md stage 12, benchmark results refreshed via osu.csv
  (results/Azusa.csv 745 ok, results/Roxy.csv 603 ok + 142 invalid)
