+++ docs/pipeline/analysis-pipeline.md
@@ -212,7 +212,7 @@ const response = await fetch(getEndpoint(), { method: "GET", cache: "no-store" }
 | 9 Ett | `withEtterna` 时 `analyzeEtternaFromText(rawText, {musicRate, scoreGoal, cvtFlag, etternaVersion})`（calc.js 现有 loader；worker 内 import.meta.url 按模块文件解析、同源 fetch 实例化 WASM，Node 侧 fs preload 不变），输出 `ettResult {values, keycount, ...}` |
 | 10 Companella 二次 Ett | Companella/Mixed && columnCount===4 && `companellaEtternaVersion ≠ etternaVersion` 时在同一 pipeline 内完成第二次 Ett（一次往返），输出 `companellaEttResult` |
 | 11 ppMetrics | `options.withPpMetrics === true` 时输出 `ppMetrics {star, variety, accScalar, totalNotes, spikiness, switches}`：Sunny/Companella 从 `selectedRework.ppMetrics` 提取；Azusa/Roxy/Mixed 从归一化用的 sunnyResult 对象提取（保留引用，同一次 Sunny 结果）；**Daniel 在 pipeline 内跑专用 Sunny pass**（`runSunnyEstimatorFromText(rawText, {...options, withPpMetrics: true}, parser)`，同 options 保证一致性，worker 内执行）；失败 → `ppMetrics: null`（软失败，不进 errors[]） |
-| 12 马拉松修正 | **恒定应用**（无设置开关）：`actualEstimatorAlgorithm ∈ {Azusa, Roxy}` 且 `ettResult.values` 可用时，对 `rework.numericDifficulty` 只降不升修正并重派生 `estDiff`（`numericToRcLabel`）：`durationS = (max−min of noteStarts)/1000`，>300s 且 MSD 技能均衡（`max(4 聚合技能)/total < 0.45`）时修正 `0.20/分钟`、封顶 `0.65`，numeric taper `10~16` 线性渐减。star 不动（归一化已覆盖）。软失败保持原结果。详见 [features/marathon-correction.md](../features/marathon-correction.md) |
+| 12 马拉松修正 | **恒定应用**（无设置开关）：`actualEstimatorAlgorithm ∈ {Azusa, Roxy}` 且 `ettResult.values` 可用时，对 `rework.numericDifficulty` 只降不升修正并重派生 `estDiff`（`numericToRcLabel`）：`durationS = (max−min of noteStarts)/1000`，>300s 且 MSD 技能均衡（`max(4 聚合技能)/total < 0.45`）时修正 `min(0.50, 0.40×ln(1+excessMin))`（对数饱和，防相邻段位课程倒挂）、numeric taper `10~16` 线性渐减。star 不动（归一化已覆盖）。软失败保持原结果。详见 [features/marathon-correction.md](../features/marathon-correction.md) |
 
 - **附属段开关**：`withPattern/withEtterna/withInterlude` 取自 needComputed（fetch 前的保守值，与缓存覆盖检查同源，analysis.js:365-368）。默认 false——仅请求需要的段，避免 worker 白算（5K 等非支持键数谱面被 override 强制 Pattern 的边界：主线程消费段有回退分支，见下）。
 - **软失败通道**：附属段各自 try/catch，失败置空字段（`patternReport=null` / `ettResult=null` / `interludeStar=NaN`）并填独立错误文本（`patternError/ettError/interludeError/companellaEttError`），**不并入 errors[]**——旧代码的 errors.push 带展示条件（`shouldReportEtternaError`/`isKeycountError` 过滤、need* 门控）依赖主线程状态，由 analysis.js 按旧条件决定是否并入，保证逐字一致。
